requirements.txt
openai
tiktoken
PyPDF2
numpy
python-dotenv